Refuse a tag no branch covers identically on both sides, and close the review
The policy tag crossed the boundary as a `u8`, so PyO3 refused anything outside
0-255 before any check of ours ran, with a `TypeError` naming neither the tag
nor the branch it lacks. The reference refused the same input as a `ValueError`
naming both. The tag is an `i64` on the Rust side now, wide enough that our own
guard is what reports it: -1, 256 and 5 are refused by both sides with the same
class and the same words.

A non-integer tag was worse than that. `2.0 in frozenset({0, 1, 2, 3, 4})` is
true, and `2.0 == KIND["risk_ranked"]` with it, so the reference ran the branch
the float happened to equal and returned a complete result while the kernel
refused to extract it at all. Both now raise `TypeError`, and the wording is
PyO3's because it refuses first.

A `policy_uniforms` of the wrong rank failed by unpacking, which produced "not
enough values to unpack (expected 2, got 1)" — precisely the fault the
per-segment guard six lines below was added to prevent, one argument over. Both
sides name the argument and the shape now.

Two more tests that could not fail: a set comparison built from the very
mapping it compared against, and a set minus itself. The first is replaced by
a check that more than one implementation was actually compared, since dropping
the kernel from the registry left it green, comparing the reference against
itself and reporting parity.

Two invariants that no runtime mutation can reach are now asserted directly,
because they fire on the edit that would break them rather than on any input:
that the rankable tags are enumerated rather than derived from the mapping, and
that the per-segment argument names and the population columns stay in step.
Both watched failing under the edit they guard — a sixth policy named in the
mapping, and a name dropped from either tuple.

The citation gate reads tracked files with undecodable bytes replaced. Walking
what git tracks means it now meets whatever is added, and a tracked image would
otherwise stop it with a decode error unrelated to citations; a replacement
character cannot spell the literal the pattern needs, so nothing is skipped.

The task plan is finished and filed, with what the kernel measured, what the
five review passes cost, and the three things left open on purpose.

Suite 246 passed, notebooks 4 passed, Rust 9 passed, sweep green through the
kernel.

+++ src/policies.rs
@@ -57,21 +57,29 @@ use pyo3::prelude::*;
 /// runs all five policies through both implementations, and any two tags
 /// exchanged funds a different set of segments.
 pub mod kind {
+    // `i64` rather than `u8`, which is wide enough for five tags but is not
+    // what decides the type here. PyO3 refuses to extract a value outside a
+    // narrow integer's range *before* any check of ours runs, and it refuses
+    // with a `TypeError` naming neither the tag nor the branch it lacks — so a
+    // tag of -1 or 256 would be reported one way by this side and another by
+    // the reference. Extracting the whole integer range and rejecting it
+    // ourselves is what lets both sides refuse the same input with the same
+    // error and the same words.
     // A module of `const` values, which is how Rust spells the namespace that
     // `policies.KIND` gets from being a dict: `kind::RISK_RANKED` here reads
     // as `KIND["risk_ranked"]` does there. A `const` is substituted at every
     // use site at compile time, so this costs nothing at run time and, unlike
     // the dict, cannot be looked up with a key that does not exist.
     /// Replaces nothing; failures are still repaired.
-    pub const RUN_TO_FAILURE: u8 = 0;
+    pub const RUN_TO_FAILURE: i64 = 0;
     /// Every segment at or over an age, oldest first.
-    pub const AGE_THRESHOLD: u8 = 1;
+    pub const AGE_THRESHOLD: i64 = 1;
     /// Failure probability weighted by what a failure would cost.
-    pub const RISK_RANKED: u8 = 2;
+    pub const RISK_RANKED: i64 = 2;
     /// Failure probability alone, ignoring consequence.
-    pub const WORST_FIRST: u8 = 3;
+    pub const WORST_FIRST: i64 = 3;
     /// A fixed per-segment priority, ignoring everything else.
-    pub const RANDOM: u8 = 4;
+    pub const RANDOM: i64 = 4;
 }
 
 /// A validated policy reduced to what the annual loop needs.
@@ -95,7 +103,7 @@ pub mod kind {
 #[derive(FromPyObject, Clone, Copy, Debug)]
 pub struct Resolved {
     /// The tag from `kind`.
-    pub kind: u8,
+    pub kind: i64,
     /// Eligibility age, compared as `age >= threshold_years`. Negative
     /// infinity makes every in-service segment eligible; positive infinity
     /// makes none, which is how `RUN_TO_FAILURE` funds nothing without a
@@ -387,7 +395,7 @@ mod tests {
     use super::*;
 
     /// A policy with the neutral threshold, so every segment is eligible.
-    fn policy(kind: u8) -> Resolved {
+    fn policy(kind: i64) -> Resolved {
         Resolved {
             kind,
             threshold_years: f64::NEG_INFINITY,
